requirements.txt
Jinja2>=3.1.2
Django>=4.1.3
django-ace>=1.0.13
django-jazzmin>=2.6.0
djangorestframework>=3.14.0
django_stubs_ext>=0.7.0
django-import-export>=3.0.1
uvicorn>=0.20.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.5
django-tinymce>=3.5.0
fontawesomefree>=6.2.1
markdown>=3.4.1
